example project created
